gps again

+++ robots/motion.go
@@ -137,7 +137,7 @@ func (m *Motion) move(p *rocproto.Packet) error {
 	lR, rR := thrust(theta, v_a, v_b)
 	lS := gobot.ToScale(gobot.FromScale(CALCSPEED*(float64(lR)/100), -90, 90), 0, 180)
 	rS := gobot.ToScale(gobot.FromScale(CALCSPEED*(float64(rR)/100), -90, 90), 0, 180)
-	n.Speed = (lR + rR) / 2
+	n.Speed = float64((lR + rR)) / 2
 	fmt.Printf("Mouvement ==> %+v\nSpeed ==> L : %v	R : %v\n", n, lS, lR)
 	gobot.Publish(m.Event("move"), *n)
 	m.motorL.Move(uint8(lS))
